added aula and avaliacao controller

+++ src/main/java/com/xicola/xicola/repository/AulaRepository.java
@@ -5,9 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AulaRepository extends JpaRepository<Aula, Long> {
+public interface AulaRepository extends JpaRepository<Aula, Integer> {
 
-
-
-
+   
 }
